Add simpler user-faced API
